removed duplicate Examples and defined Response.stream() for HTML protocol
