+++ Examples/HummingbirdDemo/Package.swift
@@ -12,7 +12,9 @@ let package = Package(
     dependencies: [
         .package(url: "https://github.com/hummingbird-project/hummingbird.git", from: "2.0.0-rc.1"),
         .package(url: "https://github.com/hummingbird-community/hummingbird-elementary.git", from: "0.1.0-rc.1"),
-        .package(path: "../../../"),
+        .package(path: "../../"),
+        .package(url: "https://github.com/apple/swift-async-algorithms", from: "1.0.0"),
+        .package(url: "https://github.com/swift-server/swift-service-lifecycle.git", from: "2.0.0"),
     ],
     targets: [
         .executableTarget(
@@ -21,6 +23,9 @@ let package = Package(
                 .product(name: "Hummingbird", package: "hummingbird"),
                 .product(name: "HummingbirdElementary", package: "hummingbird-elementary"),
                 .product(name: "ElementaryHTMX", package: "elementary-htmx"),
+                .product(name: "ElementaryHxSSE", package: "elementary-htmx"),
+                .product(name: "AsyncAlgorithms", package: "swift-async-algorithms"),
+                .product(name: "ServiceLifecycle", package: "swift-service-lifecycle"),
             ],
             resources: [
                 .copy("Public"),
