--- Examples/Vapor/VaporDemo/.gitignore
@@ -1,5 +1,0 @@
-/.build
-xcuserdata/
-DerivedData/
-.vscode
-Package.resolved
